packages/discovery/requirements.txt
absl-py==0.7.1
astor==0.7.1
bert-serving-client==1.8.9
bert-serving-server==1.8.9
certifi==2019.3.9
gast==0.2.2
GPUtil==1.4.0
grpcio==1.16.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1
mkl-fft
mkl-random
mock==2.0.0
numpy==1.16.3
pbr==5.1.3
protobuf==3.6.1
pyzmq==18.0.1
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.2
